floor atk * talent before multiplying combo

+++ battle/attack/AtkModelEntity.java
@@ -89,7 +89,7 @@ public abstract class AtkModelEntity extends AtkModelAb {
 			atks[i] = (int) (Math.round(raw[i][0] * d1) * d0);
 
 			if (pc != null && lv != null)
-				atks[i] = (int) Math.round(pc.getAtkMultiplication(lv.getLvs()) * atks[i] * (1 + ent.basis.b.getInc(Data.C_ATK) * 0.01));
+				atks[i] = (int) Math.round((int) (pc.getAtkMultiplication(lv.getLvs()) * atks[i]) * (1 + ent.basis.b.getInc(Data.C_ATK) * 0.01));
 
 			abis[i] = raw[i][2];
 			act[i] = data.getAtkModel(i).loopCount();
